+++ src/Libraries/SmartStore.Core/Domain/Seo/SeoSettings.cs
@@ -1,25 +1,68 @@
-﻿
+﻿using System;
 using System.Collections.Generic;
 using SmartStore.Core.Configuration;
 
 namespace SmartStore.Core.Domain.Seo
 {
-    public class SeoSettings : ISettings
+	public class SeoSettings : ISettings
     {
 		public SeoSettings()
 		{
 			PageTitleSeparator = ". ";
 			PageTitleSeoAdjustment = PageTitleSeoAdjustment.PagenameAfterStorename;
-			DefaultTitle = "Your store";
+			DefaultTitle = "Shop";
 			DefaultMetaKeywords = "";
 			DefaultMetaDescription = "";
-			AllowUnicodeCharsInUrls = true;
-			CanonicalHostNameRule = Seo.CanonicalHostNameRule.NoRule;
-			ReservedUrlRecordSlugs = new List<string> { "admin", "install", "recentlyviewedproducts", "newproducts", "compareproducts", "clearcomparelist", "setproductreviewhelpfulness", "login", "register", "logout", "cart", "wishlist", "emailwishlist", "checkout", "contactus", "passwordrecovery", "subscribenewsletter", "blog", "boards", "inboxupdate", "sentupdate", "news", "sitemap", "sitemapseo", "search", "config", "api", "odata" };
-            ExtraRobotsDisallows = new List<string> { "/blog/tag/", "/blog/month/", "/producttags/" };
+			AllowUnicodeCharsInUrls = false;
+			CanonicalHostNameRule = CanonicalHostNameRule.NoRule;
 			LoadAllUrlAliasesOnStartup = true;
+
+			ExtraRobotsDisallows = new List<string> { "/blog/tag/", "/blog/month/", "/producttags/" };
+
+			ReservedUrlRecordSlugs = new List<string>
+			{
+				"admin",
+				"install",
+				"recentlyviewedproducts",
+				"newproducts",
+				"compareproducts",
+				"clearcomparelist",
+				"setproductreviewhelpfulness",
+				"login",
+				"register",
+				"logout",
+				"cart",
+				"wishlist",
+				"emailwishlist",
+				"checkout",
+				"contactus",
+				"passwordrecovery",
+				"subscribenewsletter",
+				"blog",
+				"boards",
+				"inboxupdate",
+				"sentupdate",
+				"news",
+				"sitemap",
+				"sitemapseo",
+				"search",
+				"config",
+				"api",
+				"odata"
+			};
+
+			SeoNameCharConversion = string.Join(Environment.NewLine, new List<string>
+			{
+				"ä;ae",
+				"ö;oe",
+				"ü;ue",
+				"Ä;Ae",
+				"Ö;Oe",
+				"Ü;Ue",
+				"ß;ss"
+			});
 		}
-		
+
 		public string PageTitleSeparator { get; set; }
         public PageTitleSeoAdjustment PageTitleSeoAdjustment { get; set; }
         public string DefaultTitle { get; set; }
@@ -28,8 +71,9 @@ namespace SmartStore.Core.Domain.Seo
 
         public bool ConvertNonWesternChars { get; set; }
         public bool AllowUnicodeCharsInUrls { get; set; }
+		public string SeoNameCharConversion { get; set; }
 
-        public bool CanonicalUrlsEnabled { get; set; }
+		public bool CanonicalUrlsEnabled { get; set; }
 		public CanonicalHostNameRule CanonicalHostNameRule { get; set; }
 
         /// <summary>