+++ changelog.md
@@ -31,6 +31,7 @@
 * Added option to make VATIN mandatory during customer registration
 * #840 Activity log: Have option to exclude search engine activity
 * #841 Activity log for deleting an order
+* More settings to control creation of SEO names
 
 ### Improvements
 * (Perf) Implemented static caches for URL aliases and localized properties. Increases app startup and request speed by up to 30%.